requirements.txt
python-telegram-bot
PyPDF2
comtypes
pywin32
